Tests/LibWeb/Text/expected/css/flex-intrinsic-inline-block-baseline-scroll-height.txt

blank tail after spacer: 0px
